fix small bugs in hanlde failure

+++ src/mp3/application/WordCount.java
@@ -3,12 +3,12 @@ package mp3.application;
 import java.io.BufferedReader;
 import java.io.FileReader;
 
-public class WordCount extends MapleJuice<String, Integer> {
+public class WordCount extends MapleJuice<String, Long> {
     @Override
     public void maple(String line) {
         String[] infoList = line.split(" ");
         for (String info : infoList) {
-            mapleOutput.add(new Pair<>(info, 1));
+            mapleOutput.add(new Pair<>(info, 1L));
         }
     }
 
@@ -19,7 +19,7 @@ public class WordCount extends MapleJuice<String, Integer> {
             fIn = new BufferedReader(new FileReader(filePath));
             String line = fIn.readLine();
             String key = line.split(" ")[0];
-            int count = 1;
+            long count = 1;
             while((line = fIn.readLine()) != null) {
                 count++;
             }
